fix overflow of integer multiplication in graph statistics

+++ MILP-Struct/src/main/java/graph/DualGraphStatistics.java
@@ -16,7 +16,7 @@ public class DualGraphStatistics extends GraphStatistics{
         if (graphData.numNodes <= 1) {
             graphData.density = 0;
         } else {
-            graphData.density = (double) (2 * graphData.numEdges) / (double) (graphData.numNodes * (graphData.numNodes - 1));
+            graphData.density = (2 * (double) graphData.numEdges) /  ((double) graphData.numNodes * (((double) graphData.numNodes) - 1.0));
         }
     }
 }
