+++ MILP-Struct/src/main/java/graph/IncidenceGraphStatistics.java
@@ -17,6 +17,6 @@ public class IncidenceGraphStatistics extends GraphStatistics {
         // define the density for the incidence graph to be num edges / left side * right side of the bipartite graph
         // i.e. the number of edges divided by the maximum possible number of edges
         LPData lpData = lpStatistics.getLinearProgramData();
-        graphData.density = (double) (graphData.numEdges) / (double) (lpData.numConstraints * lpData.numVariables);
+        graphData.density = (double) (graphData.numEdges) / ((double) lpData.numConstraints * (double) lpData.numVariables);
     }
 }
